feat(cua-driver-rs)(recording): native ScreenCaptureKit video on macOS, ffmpeg only on Win/Linux
Replaces the macOS ffmpeg+avfoundation subprocess pipeline with an
in-process SCStream + SCRecordingOutput. ScreenCaptureKit runs in the
cua-driver process, so it inherits the daemon's Screen Recording TCC
grant — no separate subprocess prompt, no fast-fail hang heuristic.

video.rs now exposes `VideoBackend` + `VideoBackendFactory` traits;
each platform crate registers its concrete factory at startup
(`set_video_backend_factory`), mirroring how `SCREENSHOT_FN` /
`AX_SNAPSHOT_FN` are wired. macOS registers `SckitVideoBackendFactory`
(platform-macos/src/video_sckit.rs); Windows + Linux register
`FfmpegVideoBackendFactory` (cua-driver-core/src/video_ffmpeg.rs,
carrying the existing stderr-tail fast-fail).

Requires macOS 15.0+ (SCRecordingOutput introduced in macOS 15).
