+++ libs/cua-driver/rust/crates/cua-driver-core/src/lib.rs
@@ -26,5 +26,6 @@ pub mod text_sanitize;
 pub mod tool;
 pub mod tool_args;
 pub mod video;
+pub mod video_ffmpeg;
 
 pub use recording::RecordingSession;
